Unify and regroup examples
